ShopsLocationAdapter.kt was updated. Now only distance updates without redraw Recycler.

+++ app/src/main/java/com/example/coffeeapp/domain/main/shops/model/Shop.kt
@@ -6,5 +6,5 @@ data class Shop(
     val id: Long,
     val name: String,
     val point: Point,
-    val distance: Double? = null
+    val distance: Int? = null
 )
